fix(types): Unit is a type not a constant, moving to global `types` file
